All IA pages and templates working again
