+++ frontend-react/src/pages/support/Support.tsx
@@ -10,13 +10,14 @@ import {
 } from "../../components/Content/Templates/IATemplate";
 
 const templateProps: IATemplateProps<IACardGridProps> = {
-    directories: supportDirectories,
     pageName: SupportDirectoryTools.title,
     subtitle: SupportDirectoryTools.subtitle,
     templateKey: TemplateName.CARD_GRID,
     templateProps: {
         directories: supportDirectories,
     },
+    includeRouter: true,
+    directories: supportDirectories,
 };
 /** Use this component in the main App Router! It will handle rendering everything
  * and set the Helmet values */
